backend/requirements.txt
pyjwt>=2.0.0
Flask>=2.0.0
pyotp>=2.0.0
flask-cors>=3.0.0
bcrypt>=4.0.0